+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The latest *ModbusScope* installer or standalone version can always be downloade
 ### Added 
 
 - Diagnostic dialog is opened when stats section in status bar is clicked
+- Add median poll time of last 50 samples to status bar
 
 ### Fixed